Organize the project in browser and rederer
