+++ src/browser/db/sqlectron-db.js
