+++ src/renderer/actions/connections.js
