+++ src/renderer/actions/types.js
