+++ src/renderer/components/connection-list-item.jsx
