+++ src/renderer/components/connection-list.jsx
