+++ src/renderer/components/database-filter.jsx
