+++ src/renderer/components/loading.jsx
