+++ src/renderer/components/query.jsx
