+++ src/renderer/containers/query-browser.jsx
