+++ src/renderer/reducers/connections.js
