+++ src/renderer/reducers/databases.js
