+++ src/renderer/reducers/index.js
