+++ src/renderer/reducers/query.js
