+++ src/renderer/store/configure.js
